refactor: move test execution classes to separate namespace
