+++ tests/Allure.Xunit.Tests/AllureIds/AllureIdTests.cs
@@ -1,4 +1,5 @@
 using Allure.Testing;
+using Allure.Testing.Execution;
 
 namespace Allure.Xunit.Tests.AllureIds;
